Breakup config for pg and srv
Getting ready for multi-backend system, where each backend
would have its own configuration.
