+++ Cargo.toml
@@ -24,7 +24,7 @@ actix-web = "4"
 async-trait = "0.1"
 bb8 = "0.8"
 bb8-postgres = "0.8"
-clap = { version = "4.0.2", features = ["derive"] }
+clap = { version = "4", features = ["derive"] }
 env_logger = "0.9"
 itertools = "0.10"
 log = "0.4"
