+++ src/pg/composite_source.rs
@@ -1,13 +1,11 @@
-use std::io;
-
+use crate::pg::db::Connection;
+use crate::pg::table_source::TableSource;
+use crate::pg::utils::{get_bounds_cte, get_srid_bounds, prettify_error};
+use crate::source::{Query, Source, Tile, Xyz};
 use async_trait::async_trait;
 use itertools::Itertools;
+use std::io;
 use tilejson::{tilejson, Bounds, TileJSON};
-
-use crate::db::Connection;
-use crate::source::{Query, Source, Tile, Xyz};
-use crate::table_source::TableSource;
-use crate::utils::{get_bounds_cte, get_srid_bounds, prettify_error};
 
 #[derive(Clone, Debug)]
 pub struct CompositeSource {
