+++ src/pg/scripts/get_bounds_cte.sql
