+++ src/pg/scripts/get_table_sources.sql
